+++ tests/generics.rs
@@ -2,8 +2,8 @@
 #![allow(dead_code)] // some code is tested for type checking only
 
 use derive_more::{
-    Add, AddAssign, Borrow, Constructor, Deref, DerefMut, Display, Error, From,
-    FromStr, Index, IndexMut, IntoIterator, Mul, MulAssign, Not, Sum, TryInto,
+    Add, AddAssign, Borrow, BorrowMut, Constructor, Deref, DerefMut, Display, Error,
+    From, FromStr, Index, IndexMut, IntoIterator, Mul, MulAssign, Not, Sum, TryInto,
 };
 
 #[derive(
@@ -22,6 +22,7 @@ use derive_more::{
     DerefMut,
     IntoIterator,
     Borrow,
+    BorrowMut,
     Constructor
 )]
 #[deref(forward)]
@@ -56,6 +57,7 @@ struct WrappedDouble2<T: Clone, U: Clone>(T, U);
     DerefMut,
     IntoIterator,
     Borrow,
+    BorrowMut,
     Constructor
 )]
 struct WrappedWithConst<T, const C: u32>(T);
@@ -75,6 +77,7 @@ struct WrappedWithConst<T, const C: u32>(T);
     DerefMut,
     IntoIterator,
     Borrow,
+    BorrowMut,
     Constructor,
     Sum
 )]