+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   and supporting custom hash functions per field or skipping fields.
   ([#532](https://github.com/JelteF/derive_more/pull/532))
 - Add `Borrow` derive for single-field structs.
+- Add `BorrowMut` derive for single-field structs.
 
 ### Fixed